- "workoutRecordsCalculator" uses functions from "exercise-records-seed-generator". This should be refactored
